AvelarMonitor/requirements.txt
Django>=4.2.0
psutil>=5.9.0
docker>=6.0.0
requests>=2.31.0
python-decouple>=3.8
Pillow>=10.0.0
APScheduler>=3.10.0
python-dateutil>=2.8.2
pytz>=2023.3
channels[daphne]>=4.0.0
channels_redis>=4.0.0
redis>=5.0.0
websockets>=12.0
celery>=5.3.0
paramiko>=3.3.0
PyJWT>=2.8.0
